Rework and deprecate top level constants.
