+++ lib/multipart/post/parts.rb
@@ -147,3 +147,6 @@ module Multipart
     end
   end
 end
+
+Parts = Multipart::Post::Parts
+Object.deprecate_constant :Parts
