+++ lib/multipartable.rb
@@ -1,3 +1,2 @@
 warn "Top level ::Multipartable is deprecated, require 'multipart/post' and use `Multipart::Post::Multipartable` instead!"
 require_relative 'multipart/post'
-Multipartable = Multipart::Post::Multipartable
