+++ lib/parts.rb
@@ -1,3 +1,2 @@
 warn "Top level ::Parts is deprecated, require 'multipart/post' and use `Multipart::Post::Parts` instead!"
 require_relative 'multipart/post'
-Parts = Multipart::Post::Parts
